server/requirements.txt
asgiref==3.5.0
Django==3.2.12
djangorestframework==3.13.1
django-cors-headers==3.11.0
django-filter==2.4.0
pytz==2021.3
sqlparse==0.4.2
psycopg2-binary
pillow